pyproject.toml
[project]
name = "gplearn-CUDA"
dynamic = ["version"]
authors = [
  { name="Trevor Stephens", email="trev.stephens@gmail.com" },
]
description = "GPU-accelerated Genetic Programming in Python (CUDA extension of gplearn)"
readme = "README.rst"
requires-python = ">=3.11"
dependencies = [
    "scikit-learn>=1.8.0",
    "joblib>=1.3.0",
    "numpy>=1.24.0"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Topic :: Software Development",
    "Topic :: Scientific/Engineering",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14"
]
license = "BSD-3-Clause"
license-files = ["LICENSE"]

[project.optional-dependencies]
cuda = [
    "cupy-cuda12x>=13.0.0"
]

[project.urls]
homepage = "https://github.com/LGA-Personal/gplearn-CUDA"
source = "https://github.com/LGA-Personal/gplearn-CUDA"
tracker = "https://github.com/LGA-Personal/gplearn-CUDA/issues"

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gplearn"]

[tool.hatch.version]
path = "gplearn/__init__.py"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E203", "E501"]
